@@ -4,75 +4,77 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@fancyapps/ui": "^4.0.15",
-        "@fortawesome/fontawesome-free": "^5.15.4",
-        "@fullcalendar/core": "^5.10.1",
-        "@fullcalendar/daygrid": "^5.10.1",
-        "@fullcalendar/interaction": "^5.10.1",
-        "@fullcalendar/moment": "^5.10.1",
-        "@fullcalendar/react": "^5.10.1",
-        "@fullcalendar/resource-timegrid": "^5.10.1",
-        "@fullcalendar/resource-timeline": "^5.10.1",
-        "@fullcalendar/rrule": "^5.10.1",
-        "@fullcalendar/scrollgrid": "^5.10.1",
-        "@fullcalendar/timegrid": "^5.10.1",
-        "@popperjs/core": "^2.11.2",
-        "@reduxjs/toolkit": "^1.7.1",
-        "@stripe/react-stripe-js": "^1.7.1",
-        "@stripe/stripe-js": "^1.27.0",
-        "@testing-library/jest-dom": "^5.16.1",
-        "@testing-library/react": "^12.1.2",
-        "@testing-library/user-event": "^13.5.0",
-        "@typescript-eslint/eslint-plugin": "^5.11.0",
+        "@fancyapps/ui": "^4.0.27",
+        "@fortawesome/fontawesome-free": "^6.1.1",
+        "@fullcalendar/core": "^5.11.0",
+        "@fullcalendar/daygrid": "^5.11.0",
+        "@fullcalendar/interaction": "^5.11.0",
+        "@fullcalendar/moment": "^5.11.0",
+        "@fullcalendar/react": "^5.11.1",
+        "@fullcalendar/resource-timegrid": "^5.11.0",
+        "@fullcalendar/resource-timeline": "^5.11.0",
+        "@fullcalendar/rrule": "^5.11.0",
+        "@fullcalendar/scrollgrid": "^5.11.0",
+        "@fullcalendar/timegrid": "^5.11.0",
+        "@popperjs/core": "^2.11.5",
+        "@reduxjs/toolkit": "^1.8.3",
+        "@stripe/react-stripe-js": "^1.9.0",
+        "@stripe/stripe-js": "^1.32.0",
+        "@testing-library/jest-dom": "^5.16.4",
+        "@testing-library/react": "^13.3.0",
+        "@testing-library/user-event": "^14.2.1",
+        "@typescript-eslint/eslint-plugin": "^5.30.4",
         "agent-base": "^6.0.2",
-        "axios": "^0.24.0",
-        "bootstrap": "^5.1.3",
+        "autoprefixer": "^10.4.7",
+        "axios": "^0.27.2",
+        "bootstrap": "^5.2.0-beta1",
         "bootstrap-daterangepicker": "^3.1.0",
-        "feather-icons": "^4.28.0",
+        "feather-icons": "^4.29.0",
         "font-awesome": "^4.7.0",
         "formik": "^2.2.9",
-        "history": "^5.2.0",
-        "i18next": "^21.6.5",
-        "i18next-browser-languagedetector": "^6.1.2",
+        "history": "^5.3.0",
+        "i18next": "^21.8.11",
+        "i18next-browser-languagedetector": "^6.1.4",
         "i18next-xhr-backend": "^3.2.2",
         "jquery": "^3.6.0",
         "lodash": "^4.17.21",
-        "moment": "^2.29.1",
+        "moment": "^2.29.3",
         "moment-timezone": "^0.5.34",
         "moment-utl": "^2.22.0",
         "postinstall": "^0.7.4",
         "prop-types": "^15.8.1",
-        "react": "^17.0.2",
-        "react-bootstrap": "^2.1.0",
+        "react": "^18.2.0",
+        "react-bootstrap": "^2.4.0",
         "react-bootstrap-daterangepicker": "^7.0.0",
         "react-credit-cards": "^0.8.3",
-        "react-dom": "^17.0.2",
-        "react-helmet-async": "^1.2.2",
+        "react-dom": "^18.2.0",
+        "react-helmet-async": "^1.3.0",
         "react-html-table-to-excel": "^2.0.0",
-        "react-i18next": "^11.15.3",
+        "react-i18next": "^11.17.4",
         "react-infinite-scroll-component": "^6.1.0",
         "react-input-mask": "^3.0.0-alpha.2",
-        "react-lines-ellipsis": "^0.15.0",
-        "react-moment": "^1.1.1",
+        "react-lines-ellipsis": "^0.15.1",
+        "react-moment": "^1.1.2",
         "react-multi-date-picker": "^3.3.0",
-        "react-redux": "^7.2.6",
-        "react-router": "^6.2.1",
-        "react-router-dom": "^6.2.1",
+        "react-redux": "^8.0.2",
+        "react-router": "^6.3.0",
+        "react-router-dom": "^6.3.0",
         "react-router-title": "^6.0.1",
-        "react-scripts": "5.0.0",
-        "react-select": "^5.2.1",
+        "react-scripts": "5.0.1",
+        "react-select": "^5.4.0",
         "react-to-print": "^2.14.7",
-        "react-toastify": "^8.2.0",
+        "react-toastify": "^9.0.5",
         "react-toolkit": "^2.0.0-alpha.6",
-        "redux": "^4.1.2",
+        "redux": "^4.2.0",
         "redux-logger": "^3.0.6",
         "redux-persist": "^6.0.0",
-        "rrule": "^2.6.9",
-        "sweetalert2": "^11.4.4",
+        "rrule": "^2.7.0",
+        "stream-http": "^3.2.0",
+        "sweetalert2": "^11.4.19",
         "sweetalert2-react": "^0.8.3",
-        "sweetalert2-react-content": "^4.2.0",
+        "sweetalert2-react-content": "^5.0.1",
         "toastr": "^2.1.4",
-        "web-vitals": "^2.1.3",
+        "web-vitals": "^2.1.4",
         "yup": "^0.32.11"
     },
     "scripts": {
